Refactor ui/utils.js tests to utils/ui/contextualballoon.js and utils/ui/table-properties.js.
